+++ .eslintrc.yml
@@ -22,19 +22,21 @@ ignorePatterns:
   - '*.js'
   - '*.d.ts'
 rules:
-  "@typescript-eslint/naming-convention": [
+  '@typescript-eslint/naming-convention': [
     error,
     { selector: default, format: [camelCase] },
     { selector: variable, format: [camelCase, UPPER_CASE] },
     { selector: parameter, format: [camelCase], leadingUnderscore: allow },
     { selector: memberLike, modifiers: [private], format: [camelCase] },
-    { selector: typeLike, format: [PascalCase] }
+    { selector: typeLike, format: [PascalCase] },
+    { selector: function, modifiers: [global], format:  [camelCase], trailingUnderscore: allow },
+    { selector: classProperty, modifiers: [static, readonly], format:  [UPPER_CASE] }
   ]
   '@typescript-eslint/no-unused-vars': error
-  camelcase: off
+  camelcase: 'off'
   import/extensions: [error, ignorePackages, { ts: never, js: never }]
   import/prefer-default-export: off
   lines-between-class-members: [error, always, { exceptAfterSingleLine: true }]
   no-console: [error, { allow: [info, warn, error] }]
-  no-underscore-dangle: [error, { allowAfterThis: true }]
-  no-unused-vars: off
+  no-underscore-dangle: 'off'
+  no-unused-vars: 'off'
